Feature/provide endpoint for patient

+++ src/main/java/vn/mavn/patientservice/service/impl/AdvertisingSourceServiceImpl.java
@@ -1,18 +1,23 @@
 package vn.mavn.patientservice.service.impl;
 
 import java.util.Collections;
+import java.util.List;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.CollectionUtils;
 import vn.mavn.patientservice.dto.AdvertisingSourceAddDto;
 import vn.mavn.patientservice.dto.AdvertisingSourceEditDto;
+import vn.mavn.patientservice.dto.qobject.QueryAdvertisingSourceDto;
 import vn.mavn.patientservice.entity.AdvertisingSource;
+import vn.mavn.patientservice.entity.MedicalRecord;
 import vn.mavn.patientservice.exception.ConflictException;
 import vn.mavn.patientservice.exception.NotFoundException;
 import vn.mavn.patientservice.repository.AdvertisingSourceRepository;
+import vn.mavn.patientservice.repository.MedicalRecordRepository;
 import vn.mavn.patientservice.repository.spec.AdvertisingSourceSpec;
 import vn.mavn.patientservice.service.AdvertisingSourceService;
 
@@ -23,6 +28,9 @@ public class AdvertisingSourceServiceImpl implements AdvertisingSourceService {
   @Autowired
   private AdvertisingSourceRepository advertisingSourceRepository;
 
+  @Autowired
+  private MedicalRecordRepository medicalRecordRepository;
+
   @Override
   public AdvertisingSource addNew(AdvertisingSourceAddDto advertisingSourceAddDto) {
     //TODO: valid name duplicate
@@ -32,7 +40,8 @@ public class AdvertisingSourceServiceImpl implements AdvertisingSourceService {
         });
     AdvertisingSource advertisingSource = AdvertisingSource.builder()
         .description(advertisingSourceAddDto.getDescription())
-        .name(advertisingSourceAddDto.getName().trim()).build();
+        .name(advertisingSourceAddDto.getName().trim())
+        .isActive(advertisingSourceAddDto.getIsActive()).build();
     advertisingSource.setCreatedBy(advertisingSourceAddDto.getCreatedBy());
     advertisingSource.setUpdatedBy(advertisingSourceAddDto.getCreatedBy());
     return advertisingSourceRepository.save(advertisingSource);
@@ -46,7 +55,7 @@ public class AdvertisingSourceServiceImpl implements AdvertisingSourceService {
             Collections.singletonList("err-advertising-not-found")));
     //TODO: valid name duplicate
     advertisingSourceRepository
-        .findByNameNotEqualId(advertisingSourceEditDto.getName().trim().toUpperCase(),
+        .findByNameNotEqualId(advertisingSourceEditDto.getName().trim(),
             advertisingSourceEditDto.getId())
         .ifPresent(advert -> {
           throw new ConflictException(Collections.singletonList("err-advertising-duplicate-name"));
@@ -69,12 +78,20 @@ public class AdvertisingSourceServiceImpl implements AdvertisingSourceService {
     AdvertisingSource advertisingSource = advertisingSourceRepository
         .findById(id).orElseThrow(() -> new NotFoundException(
             Collections.singletonList("err-advertising-not-found")));
+    //TODO: valid advertising used or not
+    List<MedicalRecord> medicalRecords = medicalRecordRepository
+        .findByAvertId(advertisingSource.getId());
+    if (!CollectionUtils.isEmpty(medicalRecords)) {
+      throw new ConflictException(
+          Collections.singletonList("err-advertising-delete-not-successfully"));
+    }
     advertisingSourceRepository.deleteAdvert(advertisingSource.getId());
   }
 
   @Override
-  public Page<AdvertisingSource> findAll(String name, Pageable pageable) {
+  public Page<AdvertisingSource> findAll(QueryAdvertisingSourceDto queryAdvertisingSourceDto,
+      Pageable pageable) {
     return (Page<AdvertisingSource>) advertisingSourceRepository.findAll(
-        AdvertisingSourceSpec.findAllProfiles(name), pageable);
+        AdvertisingSourceSpec.findAllAdvert(queryAdvertisingSourceDto), pageable);
   }
 }
